+++ torchmetrics/metric.py
@@ -81,12 +81,32 @@ class Metric(Module, ABC):
         dist_sync_on_step:
             Synchronize metric state across processes at each ``forward()``
             before returning the value at the step.
+
+            .. deprecated:: v0.8
+                Argument is deprecated and will be removed in v0.9 in favour of instead
+                passing it in as keyword argument.
+
         process_group:
-            Specify the process group on which synchronization is called.
-            default: `None` (which selects the entire world)
+            Specify the process group on which synchronization is called. Defaults is `None`
+            which selects the entire world
+
+            .. deprecated:: v0.8
+                Argument is deprecated and will be removed in v0.9 in favour of instead
+                passing it in as keyword argument.
+
         dist_sync_fn:
             Callback that performs the allgather operation on the metric state. When `None`, DDP
             will be used to perform the allgather.
+
+            .. deprecated:: v0.8
+                Argument is deprecated and will be removed in v0.9 in favour of instead
+                passing it in as keyword argument.
+
+        kwargs: additional keyword arguments, see :ref:`Metric kwargs` for more info.
+
+            - dist_sync_on_step: If metric state should synchronize on ``forward()``
+            - process_group: The process group on which the synchronization is called
+            - dist_sync_fn: function that performs the allgather option on the metric state
     """
 
     __jit_ignored_attributes__ = ["device"]
@@ -97,9 +117,7 @@ class Metric(Module, ABC):
     def __init__(
         self,
         compute_on_step: Optional[bool] = None,
-        dist_sync_on_step: bool = False,
-        process_group: Optional[Any] = None,
-        dist_sync_fn: Callable = None,
+        **kwargs: Dict[str, Any],
     ) -> None:
         super().__init__()
 
@@ -113,13 +131,19 @@ class Metric(Module, ABC):
             warnings.warn(
                 "Argument `compute_on_step` is deprecated in v0.8 and will be removed in v0.9", DeprecationWarning
             )
+        self.dist_sync_on_step = kwargs.pop("dist_sync_on_step", False)
+        if not isinstance(self.dist_sync_on_step, bool):
+            raise ValueError(
+                f"Expected keyword argument `dist_sync_on_step` to be an `bool` but got {self.dist_sync_on_step}"
+            )
+        self.process_group = kwargs.pop("process_group", None)
+        self.dist_sync_fn = kwargs.pop("dist_sync_fn", None)
+        if self.dist_sync_fn is not None and not callable(self.dist_sync_fn):
+            raise ValueError(
+                f"Expected keyword argument `dist_sync_fn` to be an callable function but got {self.dist_sync_fn}"
+            )
 
-        self.dist_sync_on_step = dist_sync_on_step
-        self.process_group = process_group
-        self.dist_sync_fn = dist_sync_fn
-        self._to_sync = True
-        self._should_unsync = True
-
+        # initialize
         if is_overridden("update", self, Metric):
             warnings.warn(
                 "We detected that you have overwritten the ``update`` method, which was the API"
@@ -135,15 +159,19 @@ class Metric(Module, ABC):
                 DeprecationWarning,
             )
 
-        self._update_signature = inspect.signature(self._update)
+        self._update_signature = inspect.signature(self.update)
+        self.update: Callable = self._wrap_update(self.update)  # type: ignore
+        self.compute: Callable = self._wrap_compute(self.compute)  # type: ignore
         self._computed = None
         self._forward_cache = None
         self._update_called = False
+        self._to_sync = True
+        self._should_unsync = True
 
         # initialize state
         self._defaults: Dict[str, Union[List, Tensor]] = {}
         self._persistent: Dict[str, bool] = {}
-        self._reductions: Dict[str, Union[str, Callable[[Union[List[Tensor], Tensor]], Tensor], None]] = {}
+        self._reductions: Dict[str, Union[str, Callable[..., Any], None]] = {}
 
         # state management
         self._is_synced = False
@@ -234,7 +262,7 @@ class Metric(Module, ABC):
         with torch.no_grad():
             self.update(*args, **kwargs)
 
-        self._to_sync = self.dist_sync_on_step
+        self._to_sync = self.dist_sync_on_step  # type: ignore
         # skip restore cache operation from compute as cache is stored below.
         self._should_unsync = False
 
@@ -381,6 +409,7 @@ class Metric(Module, ABC):
 
         self.unsync(should_unsync=self._is_synced and should_unsync)
 
+
     def compute(self) -> Any:
         if not self._update_called:
             rank_zero_warn(
@@ -389,6 +418,32 @@ class Metric(Module, ABC):
                 " as metric states have not yet been updated.",
                 UserWarning,
             )
+
+    def _wrap_compute(self, compute: Callable) -> Callable:
+        @functools.wraps(compute)
+        def wrapped_func(*args: Any, **kwargs: Any) -> Any:
+            if not self._update_called:
+                rank_zero_warn(
+                    f"The ``compute`` method of metric {self.__class__.__name__}"
+                    " was called before the ``update`` method which may lead to errors,"
+                    " as metric states have not yet been updated.",
+                    UserWarning,
+                )
+
+            # return cached value
+            if self._computed is not None:
+                return self._computed
+
+            # compute relies on the sync context manager to gather the states across processes and apply reduction
+            # if synchronization happened, the current rank accumulated states will be restored to keep
+            # accumulation going if ``should_unsync=True``,
+            with self.sync_context(
+                dist_sync_fn=self.dist_sync_fn,  # type: ignore
+                should_sync=self._to_sync,
+                should_unsync=self._should_unsync,
+            ):
+                value = compute(*args, **kwargs)
+                self._computed = _squeeze_if_scalar(value)
 
         # return cached value
         if self._computed is not None:
@@ -482,7 +537,7 @@ class Metric(Module, ABC):
         """
         return self
 
-    def set_dtype(self, dst_type: Union[str, torch.dtype]) -> None:
+    def set_dtype(self, dst_type: Union[str, torch.dtype]) -> "Metric":
         """Special version of `type` for transferring all metric states to specific dtype
         Arguments:
             dst_type (type or string): the desired type